+++ vectors.h
@@ -21,30 +21,30 @@ typedef struct		s_screen_xy
 
 typedef struct		s_xy
 {
-	double			x;
-	double			y;
+	float			x;
+	float			y;
 }					t_xy;
 
 typedef t_xy		t_range;
 
 typedef struct		s_xyz
 {
-	double			x;
-	double			y;
-	double			z;
+	float			x;
+	float			y;
+	float			z;
 }					t_xyz;
 
 /*
 ** Vector functions
 */
 
-t_xy			vec2(double x, double y);
+t_xy			vec2(float x, float y);
 t_xy			vec2_add(t_xy a, t_xy b);
 t_xy			vec2_dec(t_xy a, t_xy b);
-t_xy			vec2_rot(t_xy v, double angle);
+t_xy			vec2_rot(t_xy v, float angle);
 t_xy			vec2_norm(t_xy v);
-t_xy			vec2_mul(t_xy v, double scalar);
-double			vec2_dot(t_xy a, t_xy b);
-double			ft_map(double in, t_range from, t_range to);
+t_xy			vec2_mul(t_xy v, float scalar);
+float			vec2_dot(t_xy a, t_xy b);
+float			ft_map(float in, t_range from, t_range to);
 
 #endif
